+++ operation.go
@@ -438,7 +438,7 @@ func (c *OperationClientImpl) ListOrders(request *ListOrderRequest, machineCode 
 type OrderGoods struct {
 	ItemCode  string `json:"itemCode"`  // Product code
 	ItemName  string `json:"itemName"`  // Product name
-	ItemPrice int    `json:"itemPrice"` // Commodity prices
+	ItemPrice float64    `json:"itemPrice"` // Commodity prices
 	Count     int    `json:"count"`     // Quantity of goods
 }
 
@@ -449,8 +449,8 @@ type OrderCallbackResponse struct {
 
 type Goods struct {
 	ItemCode      string `json:"itemCode,omitempty"`
-	ActualPrice   int    `json:"actualPrice,omitempty"`
-	OriginalPrice int    `json:"originalPrice,omitempty"`
+	ActualPrice   float64    `json:"actualPrice,omitempty"`
+	OriginalPrice float64    `json:"originalPrice,omitempty"`
 	Count         int    `json:"count,omitempty"`
 }
 
@@ -462,11 +462,11 @@ type Order struct {
 	UserCode        string  `json:"userCode,omitempty"`
 	HandleStatus    int     `json:"handleStatus,omitempty"`
 	ShopMove        int     `json:"shopMove,omitempty"`
-	TotalFee        int     `json:"totalFee,omitempty"`
+	TotalFee        float64     `json:"totalFee,omitempty"`
 	OpenDoorTime    int64   `json:"openDoorTime,omitempty"`
 	CloseDoorTime   int64   `json:"closeDoorTime,omitempty"`
-	OpenDoorWeight  int     `json:"openDoorWeight,omitempty"`
-	CloseDoorWeight int     `json:"closeDoorWeight,omitempty"`
+	OpenDoorWeight  float64     `json:"openDoorWeight,omitempty"`
+	CloseDoorWeight float64     `json:"closeDoorWeight,omitempty"`
 	OrderGoodsList  []Goods `json:"orderGoodsList,omitempty"`
 }
 
@@ -607,11 +607,11 @@ type SearchOpenDoorData struct {
 	VmCode          string  `json:"vmCode"`
 	MachineId       int     `json:"machineId"`
 	HandleStatus    int     `json:"handleStatus"`
-	TotalFee        int     `json:"totalFee"`
+	TotalFee        float64     `json:"totalFee"`
 	OpenDoorTime    int64   `json:"openDoorTime"`
 	CloseDoorTime   int64   `json:"closeDoorTime"`
-	OpenDoorWeight  int     `json:"openDoorWeight"`
-	CloseDoorWeight int     `json:"closeDoorWeight"`
+	OpenDoorWeight  float64     `json:"openDoorWeight"`
+	CloseDoorWeight float64     `json:"closeDoorWeight"`
 	OrderGoodsList  []Goods `json:"orderGoodsList"`
 	ShopMove        int     `json:"shopMove"`
 	ScanCode        string  `json:"scanCode"`
